+++ CHANGES.rst
@@ -14,6 +14,7 @@ Changelog
 - `to-alpaca` writer now has the `-a/--ensure_ascii` flag to enforce ASCII compatibility in the output
 - added global option `-u/--update_interval` to `convert` tool to customize how often progress of # records
   processed is being output in the console (default: 1000)
+- `text-length` filter now handles None values, i.e., ignores them
 
 
 0.0.3 (2023-11-10)